shui-server/pyproject.toml
[project]
name = "shui"
version = "0.1.0"
description = "A SHACL UI backend implementation."
authors = [
    {name = "Edmond Chuc", email = "edmond@kurrawong.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.5.0",
    "loguru>=0.7.2",
    "pydantic-settings>=2.4.0",
    "rdflib>=7.0.0",
    "fastapi-users[redis]>=13.0.0",
    "httpx>=0.27.2",
    "uvicorn>=0.30.6",
    "jinja2>=3.1.4",
    "dominate>=2.9.1",
    "starlette-wtf>=0.4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "pytest>=8.3.2",
]

[tool.ruff.lint]
extend-select = ["I"]
